Update UI: Subscription and TrainingClass card and Login/Signup/ResetPassword forms Backend: Remove redundant check for authenticated User

+++ Project_Spring/src/main/java/com/spring/project/service/impl/ReservationServiceImpl.java
@@ -10,16 +10,11 @@ import com.spring.project.model.User;
 import com.spring.project.model.Reservation;
 import com.spring.project.repository.ReservationRepository;
 import com.spring.project.service.ReservationService;
+import com.spring.project.util.UtilMethods;
 import lombok.AllArgsConstructor;
-import org.springframework.core.io.ClassPathResource;
-import org.springframework.core.io.Resource;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
-import org.springframework.util.StreamUtils;
-
-import java.io.IOException;
-import java.nio.charset.StandardCharsets;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
@@ -34,6 +29,7 @@ public class ReservationServiceImpl implements ReservationService {
     private final EmailSender emailSender;
     private final ReservationMapper reservationMapper;
     private final ClientService clientService;
+    private final UtilMethods utilMethods;
 
     public void saveReservation(ReservationRequest reservationRequest) {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
@@ -44,7 +40,7 @@ public class ReservationServiceImpl implements ReservationService {
                 if(reservationsForCurrentDayForUser.size() < 3) {
                     Reservation reservation = reservationMapper.convertFromDto(reservationRequest, user);
                     fotballReservationServiceImpl.save(reservation);
-                    String emailTemplate = loadEmailTemplateFromResource("reservationResponseEmail.html");
+                    String emailTemplate = utilMethods.loadEmailTemplateFromResource("reservationResponseEmail.html");
                     emailTemplate = emailTemplate.replace("${email}", authentication.getName());
                     emailTemplate = emailTemplate.replace("${hourSchedule}", reservationRequest.getHourSchedule());
                     emailTemplate = emailTemplate.replace("${dateTime}", reservationRequest.getLocalDate());
@@ -63,16 +59,6 @@ public class ReservationServiceImpl implements ReservationService {
 
     }
 
-    private String loadEmailTemplateFromResource(String fileName) {
-        try {
-            Resource resource = new ClassPathResource(fileName);
-            return StreamUtils.copyToString(resource.getInputStream(), StandardCharsets.UTF_8);
-        } catch (IOException e) {
-            e.printStackTrace();
-            return "";
-        }
-    }
-
     @Override
     public List<Reservation> getAllReservations() {
         return reservationRepository.findAllByOrderByReservationDateAsc();
@@ -87,10 +73,17 @@ public class ReservationServiceImpl implements ReservationService {
     @Override
     public void deleteReservation(Long id) {
         Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        fotballReservationServiceImpl.deleteReservation(id);
-        String emailTemplate = loadEmailTemplateFromResource("deleteReservationEmail.html");
-        emailTemplate = emailTemplate.replace("${email}", authentication.getName());
-        emailSender.send(authentication.getName(), emailTemplate, "Reservation was deleted");
+        if(authentication.isAuthenticated()) {
+            Reservation reservation = reservationRepository.findById(id).orElse(null);
+            if(reservation != null && (reservation.getReservationDate().isBefore(LocalDate.now()) || reservation.getReservationDate().isEqual(LocalDate.now()) )) {
+                fotballReservationServiceImpl.deleteReservation(id);
+                String emailTemplate = utilMethods.loadEmailTemplateFromResource("deleteReservationEmail.html");
+                emailTemplate = emailTemplate.replace("${email}", authentication.getName());
+                emailSender.send(authentication.getName(), emailTemplate, "Reservation was deleted");
+            } else {
+                throw new CreateReservationException("Reservation can not be deleted");
+            }
+        }
     }
 
     @Override
@@ -103,7 +96,7 @@ public class ReservationServiceImpl implements ReservationService {
             List<Reservation> reservations = fotballReservationServiceImpl.getReservationsByCourt(court);
             if(reservations != null) {
                 return reservations.stream()
-                        .map(courtReservation -> reservationMapper.convertToDto(courtReservation)).collect(Collectors.toList());
+                        .map(reservationMapper::convertToDto).collect(Collectors.toList());
             }else {
                 return new ArrayList<>();
             }
